Adds Sign Up Pages
# Small Changes:
Changed the action on the login form.

# Big Changes:
Adds three new pages, signupLanding, signupPatient, and signupDoctor
Patient and Doctor will ultimately only have one difference between the two of them and that is the clinicID field.
Still need to create Javascript for validation for phone numbers, email addresses, and passwords. 
Signup landing will offer two buttons to take you to the two different forms.
Still need to add form color changes based on content (or lack of content).
May want to make a separate log in page for doctors and patients.